feat: Plugin logs management in config api using  Composite Configuration
