+++ plugins/admin-ui-plugin/src/main/java/io/jans/ca/plugin/adminui/extensions/AdminUIExtension.java
@@ -1,0 +1,6 @@
+package io.jans.ca.plugin.adminui.extensions;
+
+import javax.enterprise.inject.spi.Extension;
+
+public class AdminUIExtension implements Extension {
+}
