+++ plugins/admin-ui-plugin/src/main/java/io/jans/ca/plugin/adminui/model/config/AUIConfiguration.java
@@ -1,0 +1,226 @@
+package io.jans.ca.plugin.adminui.model.config;
+
+public class AUIConfiguration {
+
+    //auth server
+    private String authServerHost;
+    private String authServerClientId;
+    private String authServerClientSecret;
+    private String authServerScope;
+    private String authServerRedirectUrl;
+    private String authServerFrontChannelLogoutUrl;
+    private String authServerPostLogoutRedirectUri;
+    private String authServerAuthzBaseUrl;
+    private String authServerTokenEndpoint;
+    private String authServerIntrospectionEndpoint;
+    private String authServerUserInfoEndpoint;
+    private String authServerEndSessionEndpoint;
+    //token server
+    private String tokenServerClientId;
+    private String tokenServerClientSecret;
+    private String tokenServerScope;
+    private String tokenServerRedirectUrl;
+    private String tokenServerFrontChannelLogoutUrl;
+    private String tokenServerPostLogoutRedirectUri;
+    private String tokenServerAuthzBaseUrl;
+    private String tokenServerTokenEndpoint;
+    private String tokenServerIntrospectionEndpoint;
+    private String tokenServerUserInfoEndpoint;
+    private String tokenServerEndSessionEndpoint;
+
+    // LicenseSpring
+    private LicenseConfiguration licenseConfiguration;
+
+    public String getAuthServerHost() {
+        return authServerHost;
+    }
+
+    public void setAuthServerHost(String authServerHost) {
+        this.authServerHost = authServerHost;
+    }
+
+    public String getAuthServerClientId() {
+        return authServerClientId;
+    }
+
+    public void setAuthServerClientId(String authServerClientId) {
+        this.authServerClientId = authServerClientId;
+    }
+
+    public String getAuthServerClientSecret() {
+        return authServerClientSecret;
+    }
+
+    public void setAuthServerClientSecret(String authServerClientSecret) {
+        this.authServerClientSecret = authServerClientSecret;
+    }
+
+    public String getAuthServerScope() {
+        return authServerScope;
+    }
+
+    public void setAuthServerScope(String authServerScope) {
+        this.authServerScope = authServerScope;
+    }
+
+    public String getAuthServerRedirectUrl() {
+        return authServerRedirectUrl;
+    }
+
+    public void setAuthServerRedirectUrl(String authServerRedirectUrl) {
+        this.authServerRedirectUrl = authServerRedirectUrl;
+    }
+
+    public String getAuthServerFrontChannelLogoutUrl() {
+        return authServerFrontChannelLogoutUrl;
+    }
+
+    public void setAuthServerFrontChannelLogoutUrl(String authServerFrontChannelLogoutUrl) {
+        this.authServerFrontChannelLogoutUrl = authServerFrontChannelLogoutUrl;
+    }
+
+    public String getAuthServerPostLogoutRedirectUri() {
+        return authServerPostLogoutRedirectUri;
+    }
+
+    public void setAuthServerPostLogoutRedirectUri(String authServerPostLogoutRedirectUri) {
+        this.authServerPostLogoutRedirectUri = authServerPostLogoutRedirectUri;
+    }
+
+    public String getAuthServerAuthzBaseUrl() {
+        return authServerAuthzBaseUrl;
+    }
+
+    public void setAuthServerAuthzBaseUrl(String authServerAuthzBaseUrl) {
+        this.authServerAuthzBaseUrl = authServerAuthzBaseUrl;
+    }
+
+    public String getAuthServerTokenEndpoint() {
+        return authServerTokenEndpoint;
+    }
+
+    public void setAuthServerTokenEndpoint(String authServerTokenEndpoint) {
+        this.authServerTokenEndpoint = authServerTokenEndpoint;
+    }
+
+    public String getAuthServerIntrospectionEndpoint() {
+        return authServerIntrospectionEndpoint;
+    }
+
+    public void setAuthServerIntrospectionEndpoint(String authServerIntrospectionEndpoint) {
+        this.authServerIntrospectionEndpoint = authServerIntrospectionEndpoint;
+    }
+
+    public String getAuthServerUserInfoEndpoint() {
+        return authServerUserInfoEndpoint;
+    }
+
+    public void setAuthServerUserInfoEndpoint(String authServerUserInfoEndpoint) {
+        this.authServerUserInfoEndpoint = authServerUserInfoEndpoint;
+    }
+
+    public String getAuthServerEndSessionEndpoint() {
+        return authServerEndSessionEndpoint;
+    }
+
+    public void setAuthServerEndSessionEndpoint(String authServerEndSessionEndpoint) {
+        this.authServerEndSessionEndpoint = authServerEndSessionEndpoint;
+    }
+
+    public String getTokenServerClientId() {
+        return tokenServerClientId;
+    }
+
+    public void setTokenServerClientId(String tokenServerClientId) {
+        this.tokenServerClientId = tokenServerClientId;
+    }
+
+    public String getTokenServerClientSecret() {
+        return tokenServerClientSecret;
+    }
+
+    public void setTokenServerClientSecret(String tokenServerClientSecret) {
+        this.tokenServerClientSecret = tokenServerClientSecret;
+    }
+
+    public String getTokenServerScope() {
+        return tokenServerScope;
+    }
+
+    public void setTokenServerScope(String tokenServerScope) {
+        this.tokenServerScope = tokenServerScope;
+    }
+
+    public String getTokenServerRedirectUrl() {
+        return tokenServerRedirectUrl;
+    }
+
+    public void setTokenServerRedirectUrl(String tokenServerRedirectUrl) {
+        this.tokenServerRedirectUrl = tokenServerRedirectUrl;
+    }
+
+    public String getTokenServerFrontChannelLogoutUrl() {
+        return tokenServerFrontChannelLogoutUrl;
+    }
+
+    public void setTokenServerFrontChannelLogoutUrl(String tokenServerFrontChannelLogoutUrl) {
+        this.tokenServerFrontChannelLogoutUrl = tokenServerFrontChannelLogoutUrl;
+    }
+
+    public String getTokenServerPostLogoutRedirectUri() {
+        return tokenServerPostLogoutRedirectUri;
+    }
+
+    public void setTokenServerPostLogoutRedirectUri(String tokenServerPostLogoutRedirectUri) {
+        this.tokenServerPostLogoutRedirectUri = tokenServerPostLogoutRedirectUri;
+    }
+
+    public String getTokenServerAuthzBaseUrl() {
+        return tokenServerAuthzBaseUrl;
+    }
+
+    public void setTokenServerAuthzBaseUrl(String tokenServerAuthzBaseUrl) {
+        this.tokenServerAuthzBaseUrl = tokenServerAuthzBaseUrl;
+    }
+
+    public String getTokenServerTokenEndpoint() {
+        return tokenServerTokenEndpoint;
+    }
+
+    public void setTokenServerTokenEndpoint(String tokenServerTokenEndpoint) {
+        this.tokenServerTokenEndpoint = tokenServerTokenEndpoint;
+    }
+
+    public String getTokenServerIntrospectionEndpoint() {
+        return tokenServerIntrospectionEndpoint;
+    }
+
+    public void setTokenServerIntrospectionEndpoint(String tokenServerIntrospectionEndpoint) {
+        this.tokenServerIntrospectionEndpoint = tokenServerIntrospectionEndpoint;
+    }
+
+    public String getTokenServerUserInfoEndpoint() {
+        return tokenServerUserInfoEndpoint;
+    }
+
+    public void setTokenServerUserInfoEndpoint(String tokenServerUserInfoEndpoint) {
+        this.tokenServerUserInfoEndpoint = tokenServerUserInfoEndpoint;
+    }
+
+    public String getTokenServerEndSessionEndpoint() {
+        return tokenServerEndSessionEndpoint;
+    }
+
+    public void setTokenServerEndSessionEndpoint(String tokenServerEndSessionEndpoint) {
+        this.tokenServerEndSessionEndpoint = tokenServerEndSessionEndpoint;
+    }
+
+    public LicenseConfiguration getLicenseConfiguration() {
+        return licenseConfiguration;
+    }
+
+    public void setLicenseConfiguration(LicenseConfiguration licenseConfiguration) {
+        this.licenseConfiguration = licenseConfiguration;
+    }
+
+}
