+++ plugins/admin-ui-plugin/src/main/java/io/jans/ca/plugin/adminui/model/exception/ApplicationException.java
@@ -1,0 +1,29 @@
+package io.jans.ca.plugin.adminui.model.exception;
+
+public class ApplicationException extends Exception {
+    private int errorCode;
+    private String message;
+
+    public ApplicationException(int errorCode, String message) {
+        super("Error Code: " + errorCode + ", Description: " + message);
+        this.errorCode = errorCode;
+        this.message = message;
+    }
+
+    public int getErrorCode() {
+        return errorCode;
+    }
+
+    public void setErrorCode(int errorCode) {
+        this.errorCode = errorCode;
+    }
+
+    @Override
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+}
