--- plugins/admin-ui-plugin/src/main/java/io/jans/configapi/plugin/adminui/extensions/AdminUIExtension.java
@@ -1,6 +1,0 @@
-package io.jans.configapi.plugin.adminui.extensions;
-
-import javax.enterprise.inject.spi.Extension;
-
-public class AdminUIExtension implements Extension {
-}
