--- plugins/admin-ui-plugin/src/main/java/io/jans/configapi/plugin/adminui/model/auth/IntrospectionResponse.java
@@ -1,140 +1,0 @@
-package io.jans.configapi.plugin.adminui.model.auth;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.util.List;
-
-public class IntrospectionResponse {
-
-    @JsonProperty("active")
-    private boolean active;
-
-    @JsonProperty("scope")
-    private List<String> scope;
-
-    @JsonProperty("client_id")
-    private String clientId;
-
-    @JsonProperty("username")
-    private String username;
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("exp")
-    private Integer expiresAt;
-
-    @JsonProperty("iat")
-    private Integer issuedAt;
-
-    @JsonProperty("sub")
-    private String subject;
-
-    @JsonProperty("aud")
-    private String audience;
-
-    @JsonProperty("iss")
-    private String issuer;
-
-    @JsonProperty("jti")
-    private String jti;
-
-    @JsonProperty("acr_values")
-    private String acrValues;
-
-    public boolean isActive() {
-        return active;
-    }
-
-    public void setActive(boolean active) {
-        this.active = active;
-    }
-
-    public List<String> getScope() {
-        return scope;
-    }
-
-    public void setScope(List<String> scope) {
-        this.scope = scope;
-    }
-
-    public String getClientId() {
-        return clientId;
-    }
-
-    public void setClientId(String clientId) {
-        this.clientId = clientId;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getTokenType() {
-        return tokenType;
-    }
-
-    public void setTokenType(String tokenType) {
-        this.tokenType = tokenType;
-    }
-
-    public Integer getExpiresAt() {
-        return expiresAt;
-    }
-
-    public void setExpiresAt(Integer expiresAt) {
-        this.expiresAt = expiresAt;
-    }
-
-    public Integer getIssuedAt() {
-        return issuedAt;
-    }
-
-    public void setIssuedAt(Integer issuedAt) {
-        this.issuedAt = issuedAt;
-    }
-
-    public String getSubject() {
-        return subject;
-    }
-
-    public void setSubject(String subject) {
-        this.subject = subject;
-    }
-
-    public String getAudience() {
-        return audience;
-    }
-
-    public void setAudience(String audience) {
-        this.audience = audience;
-    }
-
-    public String getIssuer() {
-        return issuer;
-    }
-
-    public void setIssuer(String issuer) {
-        this.issuer = issuer;
-    }
-
-    public String getJti() {
-        return jti;
-    }
-
-    public void setJti(String jti) {
-        this.jti = jti;
-    }
-
-    public String getAcrValues() {
-        return acrValues;
-    }
-
-    public void setAcrValues(String acrValues) {
-        this.acrValues = acrValues;
-    }
-}
