--- plugins/admin-ui-plugin/src/main/java/io/jans/configapi/plugin/adminui/model/auth/UserInfoRequest.java
@@ -1,28 +1,0 @@
-package io.jans.configapi.plugin.adminui.model.auth;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class UserInfoRequest {
-    private String code;
-
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public void setAccessToken(String accessToken) {
-        this.accessToken = accessToken;
-    }
-}
