+++ plugins/admin-ui-plugin/src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-io.jans.configapi.plugin.adminui.extensions.AdminUIExtension
+io.jans.ca.plugin.adminui.extensions.AdminUIExtension
