--- plugins/admin-ui-plugin/src/test/java/io/jans/configapi/plugin/adminui/TestAdminUIPluginJenkinsRunner.java
@@ -1,43 +1,0 @@
-/*
- * Janssen Project software is available under the MIT License (2008). See http://opensource.org/licenses/MIT for full text.
- *
- * Copyright (c) 2020, Janssen Project
- */
-
-package io.jans.configapi.plugin.adminui;
-
-import com.intuit.karate.Results;
-import com.intuit.karate.Runner;
-import net.masterthought.cucumber.Configuration;
-import net.masterthought.cucumber.ReportBuilder;
-import org.apache.commons.io.FileUtils;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.io.File;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-/**
- * @author Yuriy Zabrovarnyy
- */
-public class TestAdminUIPluginJenkinsRunner {
-
-    @Test
-    public void testParallel() {
-        System.setProperty("karate.env", "jenkins");
-        Results results = Runner.path("src/test/resources/feature").tags("~@ignore").parallel(1);
-        generateReport(results.getReportDir());
-        Assertions.assertEquals(0, results.getFailCount(), results.getErrorMessages());
-    }
-
-    public static void generateReport(String karateOutputPath) {
-        Collection<File> jsonFiles = FileUtils.listFiles(new File(karateOutputPath), new String[] { "json" }, true);
-        List<String> jsonPaths = new ArrayList(jsonFiles.size());
-        jsonFiles.forEach(file -> jsonPaths.add(file.getAbsolutePath()));
-        Configuration config = new Configuration(new File("target"), "karateTesting");
-        ReportBuilder reportBuilder = new ReportBuilder(jsonPaths, config);
-        reportBuilder.generateReports();
-    }
-}
